tg-bot/backend/api/requirements.txt
langchain-huggingface
langchain-qdrant
qdrant-client==1.15.0
langchain-tavily
langgraph
duckduckgo-search==8.1.1
httpx[socks]
langchain-core
langchain_community
fastapi>=0.68.0
uvicorn>=0.15.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
python-dotenv>=0.19.0
sentence_transformers
# llama-cpp-python
langchain_ollama
langchain_openai

